Adicionada a funcionalidade de envio de produtos com descontos, e selecao direta de email atraves do cliente

+++ src/main/java/br/ueg/cons/soft/estoqfacil/dto/ProdutoDTO.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 
 import java.awt.*;
 import java.io.File;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -25,9 +26,9 @@ public class ProdutoDTO {
 
     private Long quantidade;
 
-    private Double preco;
+    private BigDecimal preco;
 
-    private Double custo;
+    private BigDecimal custo;
 
     private Long imagemId;
